+++ src/test/java/old/EvanMaiTest.java
@@ -1,56 +1,30 @@
 package old;
 
 import jdk.jfr.Description;
-import net.bytebuddy.TypeCache;
+import old.runnerOld.BaseTestOld;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.openqa.selenium.chrome.ChromeOptions;
-import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.Select;
 import org.openqa.selenium.support.ui.WebDriverWait;
 import org.testng.Assert;
-import org.testng.annotations.AfterMethod;
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
-import runner.BaseTest;
 
-import javax.swing.*;
 import java.time.Duration;
 import java.util.List;
-@Ignore
-public class EvanMaiTest extends BaseTest {
 
-    private WebDriver driver;
-
-    @BeforeMethod
-    private void beforeMethod() {
-        ChromeOptions chromeOptions = new ChromeOptions();
-        chromeOptions.addArguments("--headless", "--window-size=1920,1080");
-        driver = new ChromeDriver(chromeOptions);
-        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(3));
-        getDriver().get("https://askomdch.com/");
-    }
-
-    @AfterMethod
-    private void afterMethod(){
-        driver.quit();
-    }
-
-    protected WebDriver getDriver() {
-        return driver;
-    }
+public class EvanMaiTest extends BaseTestOld {
 
     @Test
     public void testSaleIconPresent() {
+        getDriver().get("https://askomdch.com/");
+
         List<WebElement> listFeaturedProducts = getDriver().findElements(By.className("astra-shop-thumbnail-wrap"));
         List<WebElement> listFeaturedProductsText = getDriver().findElements(By.className("astra-shop-summary-wrap"));
         for (int i = 0; i < listFeaturedProductsText.size(); i++) {
             String[] arrProductText = listFeaturedProductsText.get(i).getText().split("\n");
             if (arrProductText[2].split(" ").length > 1) {
+
                 Assert.assertEquals(listFeaturedProducts.get(i).getText(), "Sale!");
             }
         }
@@ -58,6 +32,8 @@ public class EvanMaiTest extends BaseTest {
 
     @Test
     public void testDollarSignDisplayed() {
+        getDriver().get("https://askomdch.com/");
+
         List<WebElement> listFeaturedProductsText = getDriver().findElements(By.className("astra-shop-summary-wrap"));
 
         for (WebElement webElement : listFeaturedProductsText) {
@@ -69,6 +45,8 @@ public class EvanMaiTest extends BaseTest {
 
     @Test
     public void testVerifyDiscountMessagePresented() {
+        getDriver().get("https://askomdch.com/");
+
         WebElement text = getDriver().findElement(By.xpath("//h3[contains(text(), '25%')]"));
 
         Assert.assertEquals(text.getText(), "25% OFF On all products");
@@ -76,6 +54,8 @@ public class EvanMaiTest extends BaseTest {
 
     @Test
     public void testVerifyLoginPageOpened() {
+        getDriver().get("https://askomdch.com/");
+
         getDriver().findElement(By.linkText("Account")).click();
 
         Assert.assertEquals(getDriver().findElement(By.className("has-text-align-center")).getText(), "Account");
@@ -83,6 +63,8 @@ public class EvanMaiTest extends BaseTest {
 
     @Test
     public void testVerifyStorePageOpened() {
+        getDriver().get("https://askomdch.com/");
+
         getDriver().findElement(By.cssSelector("a[href='/store']")).click();
 
         WebElement store = getDriver().findElement(By.xpath("//header[contains(@class, 'woocommerce-products-header')]//following::h1"));
@@ -92,19 +74,28 @@ public class EvanMaiTest extends BaseTest {
 
     @Test
     public void testProductsDisplayed() {
+        getDriver().get("https://askomdch.com/");
+
         List<WebElement> list = getDriver().findElements(By.className("astra-shop-thumbnail-wrap"));
+
         Assert.assertEquals(list.size(), 5);
     }
 
     @Test
     public void testBackgroundColor() {
+        getDriver().get("https://askomdch.com/");
+
         getDriver().findElement(By.cssSelector("a[href='https://askomdch.com/product-category/men/']")).click();
+
         WebElement buttonSearch = getDriver().findElement(By.cssSelector("button[value='Search']"));
+
         Assert.assertEquals(buttonSearch.getCssValue("background-color"), "rgba(49, 151, 214, 1)");
     }
 
     @Test
     public void testVerifyButtonRedirect() {
+        getDriver().get("https://askomdch.com/");
+
         getDriver().findElement(By.xpath("//a[text() ='Find More']")).click();
 
         Assert.assertEquals(getDriver().findElement(By.xpath("//h1[text()='Contact Us']")).getText(), "Contact Us");
@@ -112,6 +103,8 @@ public class EvanMaiTest extends BaseTest {
 
     @Test
     public void testVerifyNumberProducts() {
+        getDriver().get("https://askomdch.com/");
+
         getDriver().findElement(By.cssSelector("a[href='https://askomdch.com/product-category/women/'")).click();
 
         List<WebElement> listOfProducts = getDriver().findElements(By.className("astra-shop-thumbnail-wrap"));
@@ -121,6 +114,8 @@ public class EvanMaiTest extends BaseTest {
 
     @Test
     public void testVerifySortByPrice() {
+        getDriver().get("https://askomdch.com/");
+
         getDriver().findElement(By.cssSelector("a[href='/store']")).click();
 
         Select drpSortBy = new Select(getDriver().findElement(By.cssSelector("select[name='orderby']")));
@@ -142,6 +137,8 @@ public class EvanMaiTest extends BaseTest {
     @Description("Verify number on shopping cart")
     @Test
     public void testVerifyNumberShoppingCart() {
+        getDriver().get("https://askomdch.com/");
+
         List<WebElement> listOfProducts = getDriver().findElements(By.xpath("//a[@data-product_id]"));
         listOfProducts.remove(4);
