+++ src/test/java/old/runnerOld/BaseTestOld.java
@@ -1,0 +1,31 @@
+package old.runnerOld;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.testng.annotations.AfterMethod;
+import org.testng.annotations.BeforeMethod;
+
+import java.time.Duration;
+
+public abstract class BaseTestOld {
+
+    private WebDriver driver;
+
+    @BeforeMethod
+    protected void beforeMethod() {
+        ChromeOptions chromeOptions = new ChromeOptions();
+        chromeOptions.addArguments("--headless", "--window-size=1920,1080");
+        driver = new ChromeDriver(chromeOptions);
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(3));
+    }
+
+    @AfterMethod
+    protected void afterMethod(){
+        driver.quit();
+    }
+
+    protected WebDriver getDriver() {
+        return driver;
+    }
+}
